+++ internal/controlplane/handlers_profile_test.go
@@ -21,11 +21,11 @@ import (
 	"github.com/mindersec/minder/internal/db/embedded"
 	"github.com/mindersec/minder/internal/engine/engcontext"
 	stubeventer "github.com/mindersec/minder/internal/events/stubs"
-	"github.com/mindersec/minder/internal/profiles"
 	"github.com/mindersec/minder/internal/providers"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
 	"github.com/mindersec/minder/pkg/engine/selectors"
+	"github.com/mindersec/minder/pkg/profiles"
 )
 
 //nolint:gocyclo
