+++ internal/engine/eval_status.go
@@ -16,7 +16,7 @@ import (
 	"github.com/mindersec/minder/internal/engine/entities"
 	evalerrors "github.com/mindersec/minder/internal/engine/errors"
 	engif "github.com/mindersec/minder/internal/engine/interfaces"
-	"github.com/mindersec/minder/internal/profiles/models"
+	"github.com/mindersec/minder/pkg/profiles/models"
 )
 
 func (e *executor) createEvalStatusParams(
