+++ internal/logger/telemetry_store_test.go
@@ -18,7 +18,7 @@ import (
 	enginerr "github.com/mindersec/minder/internal/engine/errors"
 	engif "github.com/mindersec/minder/internal/engine/interfaces"
 	"github.com/mindersec/minder/internal/logger"
-	"github.com/mindersec/minder/internal/profiles/models"
+	"github.com/mindersec/minder/pkg/profiles/models"
 )
 
 func TestTelemetryStore_Record(t *testing.T) {
