+++ pkg/profiles/rule_validator.go
