+++ pkg/profiles/store.go
@@ -11,7 +11,7 @@ import (
 	"golang.org/x/exp/maps"
 
 	"github.com/mindersec/minder/internal/db"
-	"github.com/mindersec/minder/internal/profiles/models"
+	"github.com/mindersec/minder/pkg/profiles/models"
 )
 
 // ProfileStore encapsulates operations for fetching ProfileAggregates
