+++ pkg/profiles/validator_test.go
@@ -17,9 +17,9 @@ import (
 
 	mockdb "github.com/mindersec/minder/database/mock"
 	"github.com/mindersec/minder/internal/db"
-	"github.com/mindersec/minder/internal/profiles"
 	"github.com/mindersec/minder/internal/util"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/profiles"
 )
 
 func TestValidatorScenarios(t *testing.T) {
